+++ .github/workflows/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## What's changed
 
-- Include basic info on PHP -> Frameworks & Tools -> Laravel option
+- Add docker user config